RTP-3368: kill child processes of a test on timeout
Child processes, of a test in a test suite of Lustre, are not killed
when the test is killed by Xperior on timeout.
This patch kills child processes by first getting all the child pids
and storing them. Then sending TERM signal and checking if any pid
is left. If any pid remains alive then it send KILL and return status.
Unit tests are also added for each functions added in the patch.
Change-Id: I49e893f53e4d344b8f36732e5731c3b603cbcc6b
Signed-off-by: Ashish Maurya <ashish.maurya@seagate.com>